create postsessionredirect endpoint for session checking
